refactor: repository가 너무복잡하여 domain별로 패키지 분류
