+++ src/main/java/com/team/heyyo/group/study/repository/groupstudylike/GroupStudyLikeRepositoryCustom.java
@@ -1,6 +1,7 @@
-package com.team.heyyo.group.study.repository;
+package com.team.heyyo.group.study.repository.groupstudylike;
 
 import com.team.heyyo.group.study.domain.GroupStudyLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupStudyLikeRepositoryCustom {}
+public interface GroupStudyLikeRepositoryCustom {
+}
