+++ src/main/java/com/team/heyyo/group/study/repository/groupstudylike/GroupStudyLikeRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.team.heyyo.group.study.repository;
+package com.team.heyyo.group.study.repository.groupstudylike;
 
 public class GroupStudyLikeRepositoryImpl implements GroupStudyLikeRepositoryCustom {
 }
