+++ src/main/java/com/team/heyyo/group/study/repository/groupstudytag/GroupStudyTagRepository.java
@@ -1,11 +1,13 @@
-package com.team.heyyo.group.study.repository;
+package com.team.heyyo.group.study.repository.groupstudytag;
 
 import com.team.heyyo.group.study.domain.GroupStudyTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GroupStudyTagRepository extends JpaRepository<GroupStudyTag, Long> {
+public interface GroupStudyTagRepository extends JpaRepository<GroupStudyTag, Long>, GroupStudyTagRepositoryCustom {
 
     List<GroupStudyTag> findByGroupStudyId(Long groupStudyId);
+
 }
+
